@@ -1,13 +1,22 @@
-//go:build tools
-// +build tools
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
+//go:build generate
 
 package tools
 
-// Including Terraform documentation generator following the
-// ["tools as dependencies"](https://github.com/go-modules-by-example/index/blob/master/010_tools/README.md)
-// model.
-
 import (
-	// [tfplugindocs](https://github.com/hashicorp/terraform-plugin-docs).
+	// document generation
 	_ "github.com/hashicorp/terraform-plugin-docs/cmd/tfplugindocs"
+	// copywrite header generation
+	_ "github.com/hashicorp/copywrite"
 )
+
+// Generate copyright headers
+//go:generate go run github.com/hashicorp/copywrite headers -d .. --config ../.copywrite.hcl
+// Format Terraform code for use in documentation.
+// If you do not have Terraform installed, you can remove the formatting command, but it is suggested
+// to ensure the documentation is formatted properly.
+//go:generate terraform fmt -recursive ../examples/
+// Generate documentation.
+//go:generate go run github.com/hashicorp/terraform-plugin-docs/cmd/tfplugindocs generate --provider-dir ..
