@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package provider
 
 import (
@@ -166,7 +169,7 @@ func TestAccTimeStatic_Validators(t *testing.T) {
 		Steps: []resource.TestStep{
 			{
 				Config:      testAccConfigTimeStaticRfc3339(timestamp.Format(time.RFC850)),
-				ExpectError: regexp.MustCompile(`.*must be a string in RFC3339 format`),
+				ExpectError: regexp.MustCompile(`.*Invalid RFC3339 String Value`),
 			},
 		},
 	})
